修改BUG

+++ src/main/java/test/stone/communication/entity/message/DeviceInfoMsg.java
@@ -20,7 +20,7 @@ public class DeviceInfoMsg extends TestMessage {
 
     @Getter
     @Setter
-    private short port;
+    private int port;
 
     public DeviceInfoMsg(){
         super(DEVICE_INFO);
